app-server: split transport into dedicated crate
