+++ codex-rs/app-server-transport/src/transport/auth.rs
