+++ codex-rs/app-server-transport/src/transport/remote_control/segment_tests.rs
